Config/requirements.txt
Flask==1.0.2
Flask-Compress==1.4.0
Flask-Cors==3.0.7
Flask-SSLify==0.1.5
numpy==1.16.2
pandas==0.24.1
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.3
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
mysql-connector-python
